+++ Paradigm/src/utopia/paradigm/angular/Rotation.scala
@@ -2,7 +2,8 @@ package utopia.paradigm.angular
 
 import utopia.flow.generic.model.immutable.Value
 import utopia.flow.generic.model.template.ValueConvertible
-import utopia.flow.operator.{ApproxEquals, Combinable, LinearScalable, SelfComparable, CanBeZero}
+import utopia.flow.operator.Combinable.SelfCombinable
+import utopia.flow.operator.{ApproxEquals, CanBeZero, LinearScalable, SelfComparable}
 import utopia.flow.operator.EqualsExtensions._
 import utopia.paradigm.enumeration.RotationDirection
 import utopia.paradigm.enumeration.RotationDirection.{Clockwise, Counterclockwise}
@@ -106,7 +107,7 @@ object Rotation
   * @param direction The rotation direction (default = clockwise)
 **/
 case class Rotation private(radians: Double, direction: RotationDirection = Clockwise)
-	extends LinearScalable[Rotation] with Combinable[Rotation, Rotation]
+	extends LinearScalable[Rotation] with SelfCombinable[Rotation]
 		with CanBeZero[Rotation] with SelfComparable[Rotation] with ApproxEquals[Rotation]
 		with ValueConvertible
 {