+++ Paradigm/src/utopia/paradigm/path/BezierPath.scala
@@ -1,6 +1,6 @@
 package utopia.paradigm.path
 
-import utopia.flow.operator.{Combinable, LinearMeasurable, LinearScalable}
+import utopia.flow.operator.{Combinable, HasLength, LinearScalable}
 import utopia.flow.collection.CollectionExtensions._
 
 /**
@@ -18,7 +18,7 @@ object BezierPath
 	  * @return A bezier path between paths
 	  * @throws IllegalArgumentException If points is empty
 	  */
-	def apply[P <: Combinable[P, P] with LinearScalable[P] with LinearMeasurable]
+	def apply[P <: Combinable[P, P] with LinearScalable[P] with HasLength]
 	(points: Seq[P], sequencesPerPart: Int = 6) =
 	{
 		if (points.isEmpty)
@@ -44,7 +44,7 @@ object BezierPath
 	 * @return Sequence of paths between speified points
 	 * @throws IllegalArgumentException If there are less than 3 points
 	 */
-	def parts[P <: Combinable[P, P] with LinearScalable[P] with LinearMeasurable](points: Seq[P]) =
+	def parts[P <: Combinable[P, P] with LinearScalable[P] with HasLength](points: Seq[P]) =
 	{
 		if (points.size < 3)
 			throw new IllegalArgumentException("Cubic Bezier path must be initialized with at least 3 points")
@@ -52,7 +52,7 @@ object BezierPath
 			calculatePath(points)
 	}
 	
-	private def calculatePath[P <: Combinable[P, P] with LinearScalable[P] with LinearMeasurable](points: Seq[P]) =
+	private def calculatePath[P <: Combinable[P, P] with LinearScalable[P] with HasLength](points: Seq[P]) =
 	{
 		// Number of curves
 		val n = points.size - 1